+++ src/__test__/core/accessors/AccessorSpec.ts
@@ -63,6 +63,18 @@ describe("Accessor", ()=> {
     })
     expect(this.accessor.getAggregations(["tags", "buckets"], []))
       .toEqual([1,2,3])
+
+    expect(this.accessor.getAggregations(["tags", undefined, "buckets"], []))
+      .toEqual([1,2,3])
+
+  })
+
+  it("beforeBuildQuery()", ()=> {
+    expect(this.accessor.beforeBuildQuery).toEqual(
+      jasmine.any(Function)
+    )
+    expect(this.accessor.beforeBuildQuery()).toEqual(undefined)
+
   })
 
   it("buildSharedQuery()", ()=> {