+++ src/components/search/filters/range-filter/src/RangeFilter.tsx
@@ -9,7 +9,8 @@ import {
 	RangeAccessor,
 	RenderComponentType,
 	RenderComponentPropType,
-	renderComponent
+	renderComponent,
+	FieldOptions
 } from "../../../../../core"
 
 import {
@@ -18,15 +19,8 @@ import {
 } from "../../../../ui"
 
 const defaults = require("lodash/defaults")
-const max = require("lodash/max")
-const maxBy = require("lodash/maxBy")
 const map = require("lodash/map")
 const get = require("lodash/get")
-
-function computeMaxValue(items, field) {
-  if (!items || items.length == 0) return 0
-  return maxBy(items, field)[field]
-}
 
 export interface RangeFilterProps extends SearchkitComponentProps {
 	field:string
@@ -38,7 +32,9 @@ export interface RangeFilterProps extends SearchkitComponentProps {
 	showHistogram?:boolean
 	containerComponent?: RenderComponentType<any>
   rangeComponent?: RenderComponentType<RangeProps>
-  collapsable?: boolean
+	rangeFormatter?:(count:number)=> number | string
+	marks?:Object
+	fieldOptions?:FieldOptions
 }
 
 
@@ -51,7 +47,13 @@ export class RangeFilter extends SearchkitComponent<RangeFilterProps, any> {
 		title:React.PropTypes.string.isRequired,
 		id:React.PropTypes.string.isRequired,
 		containerComponent:RenderComponentPropType,
-		rangeComponent:RenderComponentPropType
+		rangeComponent:RenderComponentPropType,
+		fieldOptions:React.PropTypes.shape({
+			type:React.PropTypes.oneOf(["embedded", "nested", "children"]).isRequired,
+			options:React.PropTypes.object
+		}),
+		rangeFormatter:React.PropTypes.func,
+		marks:React.PropTypes.object
 	}, SearchkitComponent.propTypes)
 
 
@@ -69,10 +71,11 @@ export class RangeFilter extends SearchkitComponent<RangeFilterProps, any> {
 	}
 
 	defineAccessor() {
-		const { id, title, min, max, field,
+		const { id, title, min, max, field, fieldOptions,
 			interval, showHistogram } = this.props
 		return new RangeAccessor(id,{
-			id, min, max, title, field, interval, loadBuckets:showHistogram
+			id, min, max, title, field,
+			interval, loadHistogram:showHistogram, fieldOptions
 		})
 	}
 
@@ -98,11 +101,6 @@ export class RangeFilter extends SearchkitComponent<RangeFilterProps, any> {
 		this.searchkit.performSearch()
 	}
 
-	getMaxValue() {
-		if (this.accessor.getBuckets() == 0) return 0
-		return max(map(this.accessor.getBuckets(), "doc_count"))
-	}
-
 	getRangeComponent():RenderComponentType<any>{
 	  const { rangeComponent, showHistogram } = this.props
 	  if (!showHistogram && (rangeComponent === RangeSliderHistogram)) {
@@ -113,19 +111,17 @@ export class RangeFilter extends SearchkitComponent<RangeFilterProps, any> {
 	}
 
 	render() {
-    const { id, title, containerComponent, collapsable } = this.props
-
-    const maxValue = computeMaxValue(this.accessor.getBuckets(), "doc_count")
+    const { id, title, containerComponent } = this.props
 
     return renderComponent(containerComponent, {
       title,
       className: id ? `filter--${id}` : undefined,
-      disabled: maxValue == 0
+      disabled: this.accessor.isDisabled()
     }, this.renderRangeComponent(this.getRangeComponent()))
   }
 
   renderRangeComponent(component: RenderComponentType<any>) {
-    const { min, max } = this.props
+    const { min, max, rangeFormatter, marks } = this.props
     const state = this.accessor.state.getValue()
     return renderComponent(component, {
       min, max,
@@ -133,7 +129,8 @@ export class RangeFilter extends SearchkitComponent<RangeFilterProps, any> {
       maxValue: Number(get(state, "max", max)),
       items: this.accessor.getBuckets(),
       onChange: this.sliderUpdate,
-      onFinished: this.sliderUpdateAndSearch
+      onFinished: this.sliderUpdateAndSearch,
+			rangeFormatter, marks
     })
   }