+++ src/components/search/search-box/SearchBox.tsx
@@ -4,23 +4,27 @@ import {
   QueryAccessor,
   SearchkitComponent,
   SearchkitComponentProps
-} from "../../../../core"
+} from "../../core"
 
 const defaults = require("lodash/defaults")
 const throttle = require("lodash/throttle")
 const assign = require("lodash/assign")
+const isUndefined = require("lodash/isUndefined")
 
 export interface SearchBoxProps extends SearchkitComponentProps {
   searchOnChange?:boolean
   searchThrottleTime?:number
   queryFields?:Array<string>
-  autofocus?:boolean
+  queryBuilder?:Function
   queryOptions?:any
   prefixQueryOptions?:any
+  autofocus?:boolean
   id?: string
   mod?: string
   placeholder?: string
   prefixQueryFields?:Array<string>
+  prefixQueryOptions?:Object
+  blurAction?:"search"|"restore"
 }
 
 export class SearchBox extends SearchkitComponent<SearchBoxProps, any> {
@@ -36,29 +40,34 @@ export class SearchBox extends SearchkitComponent<SearchBoxProps, any> {
   static defaultProps = {
     id: 'q',
     mod: 'sk-search-box',
-    searchThrottleTime:200
+    searchThrottleTime:200,
+    blurAction: "search"
   }
 
   static propTypes = defaults({
     id:React.PropTypes.string,
     searchOnChange:React.PropTypes.bool,
     searchThrottleTime:React.PropTypes.number,
+    queryBuilder:React.PropTypes.func,
     queryFields:React.PropTypes.arrayOf(React.PropTypes.string),
     autofocus:React.PropTypes.bool,
     queryOptions:React.PropTypes.object,
     prefixQueryOptions:React.PropTypes.object,
     prefixQueryFields:React.PropTypes.arrayOf(React.PropTypes.string),
+    prefixQueryOptions:React.PropTypes.object,
     translations:SearchkitComponent.translationsPropType(
       SearchBox.translations
     ),
     mod: React.PropTypes.string,
-    placeholder: React.PropTypes.string
+    placeholder: React.PropTypes.string,
+    blurAction: React.PropTypes.string
   }, SearchkitComponent.propTypes)
 
   constructor (props:SearchBoxProps) {
     super(props);
     this.state = {
-      focused:false
+      focused:false,
+      input: undefined
     }
     this.lastSearchMs = 0
     this.throttledSearch = throttle(()=> {
@@ -66,22 +75,27 @@ export class SearchBox extends SearchkitComponent<SearchBoxProps, any> {
     }, props.searchThrottleTime)
   }
 
-  componentWillMount() {
-    super.componentWillMount()
-  }
 
   defineBEMBlocks() {
     return { container:this.props.mod };
   }
 
   defineAccessor(){
-    const { id, prefixQueryFields, queryFields, searchOnChange, queryOptions, prefixQueryOptions } = this.props
+    const {
+      id, prefixQueryFields, queryFields, queryBuilder,
+      searchOnChange, queryOptions, prefixQueryOptions
+    } = this.props
     return new QueryAccessor(id, {
-      prefixQueryFields:prefixQueryFields || queryFields,
+      prefixQueryFields,
+      prefixQueryOptions:assign({}, prefixQueryOptions),
       queryFields:queryFields || ["_all"],
-      queryOptions:assign({
-      }, queryOptions),
-      prefixQueryOptions:assign({}, prefixQueryOptions)
+      queryOptions:assign({}, queryOptions),
+      queryBuilder,
+      onQueryStateChange: () => {
+        if (!this.unmounted && this.state.input){
+          this.setState({input: undefined})
+        }
+      }
     })
   }
 
@@ -100,20 +114,44 @@ export class SearchBox extends SearchkitComponent<SearchBoxProps, any> {
   }
 
   getValue(){
+    const { input } = this.state
+    if (isUndefined(input)) {
+      return this.getAccessorValue()
+    } else {
+      return input
+    }
+  }
+  
+  getAccessorValue(){
     return (this.accessor.state.getValue() || "") + ""
   }
 
   onChange(e){
     const query = e.target.value;
-    this.accessor.setQueryString(query)
     if (this.props.searchOnChange) {
+      this.accessor.setQueryString(query)
       this.throttledSearch()
+      this.forceUpdate()
+    } else {
+      this.setState({ input: query })
     }
-    this.forceUpdate()
   }
 
   setFocusState(focused:boolean) {
-    this.setState({focused:focused})
+    if (!focused){
+      const { input } = this.state
+      if (this.props.blurAction == "search"
+        && !isUndefined(input) 
+        && input != this.getAccessorValue()){
+        this.searchQuery(input)
+      }
+      this.setState({ 
+        focused,
+        input: undefined // Flush (should use accessor's state now)
+      })
+    } else {
+      this.setState({ focused })
+    }
   }
 
   render() {