+++ src/components/search/search-box/Searchbox.unit.tsx
@@ -1,29 +1,37 @@
 import * as React from "react";
 import {mount} from "enzyme";
-import {SearchBox} from "../src/SearchBox.tsx";
-import {SearchkitManager } from "../../../../core";
+import {SearchBox} from "./SearchBox.tsx";
+import {SearchkitManager, QueryString } from "../../../core";
 const bem = require("bem-cn");
 import {
   fastClick, hasClass, jsxToHTML, printPrettyHtml
-} from "../../../__test__/TestHelpers"
+} from "../../__test__/TestHelpers"
 
 import * as sinon from "sinon";
+
+const omit = require("lodash/omit")
 
 describe("Searchbox tests", () => {
 
   beforeEach(() => {
 
     this.searchkit = SearchkitManager.mock()
-
+    spyOn(this.searchkit, "performSearch")
     this.searchkit.translateFunction = (key)=> {
       return {
         "searchbox.placeholder":"search movies",
       }[key]
     }
 
-    this.createWrapper = (searchOnChange=false, queryFields=null, prefixQueryFields=null) => {
+    this.createWrapper = (searchOnChange=false, queryFields=null, prefixQueryFields=null, options={}) => {
       this.wrapper = mount(
-        <SearchBox searchkit={this.searchkit} searchOnChange={searchOnChange} queryFields={queryFields} prefixQueryFields={prefixQueryFields}/>
+        <SearchBox searchkit={this.searchkit}
+          searchOnChange={searchOnChange}
+          queryFields={queryFields} prefixQueryFields={prefixQueryFields}
+          queryOptions={{minimum_should_match:"60%"}}
+          prefixQueryOptions={{minimum_should_match:"70%"}}
+          {...options}
+        />
       );
       this.accessor = this.searchkit.accessors.getAccessors()[0]
     }
@@ -82,9 +90,10 @@ describe("Searchbox tests", () => {
     this.createWrapper(false)
     this.typeSearch('m')
     this.typeSearch('ma')
-    expect(this.accessor.state.getValue()).toBe("ma")
+    expect(this.accessor.state.getValue()).toBe(null)
     expect(spy.callCount).toBe(0)
     this.wrapper.find("form").simulate("submit")
+    expect(this.accessor.state.getValue()).toBe("ma")
     expect(spy.callCount).toBe(1)
   })
 
@@ -93,10 +102,12 @@ describe("Searchbox tests", () => {
 
     expect(this.accessor.key).toBe("q")
     let options = this.accessor.options
-    expect(options).toEqual({
-      "queryFields": ["title"],
-      prefixQueryFields:false,
-      "queryOptions": {}
+    expect(omit(options, "onQueryStateChange")).toEqual({
+      queryFields: ["title"],
+      prefixQueryFields:null,
+      queryOptions: {minimum_should_match:"60%"},
+      prefixQueryOptions: {minimum_should_match:"70%"},
+      queryBuilder:undefined
     })
 
   })
@@ -107,22 +118,28 @@ describe("Searchbox tests", () => {
     expect(this.accessor.key).toBe("q")
     let options = this.accessor.options
     expect(options).toEqual({
-      "queryFields": ["title"],
-      prefixQueryFields:["title"],
-      "queryOptions": {}
+      queryFields: ["title"],
+      prefixQueryFields:null,
+      queryOptions: {minimum_should_match:"60%"},
+      prefixQueryOptions: {minimum_should_match:"70%"},
+      queryBuilder:undefined,
+      onQueryStateChange:jasmine.any(Function)
     })
 
   })
 
   it("should configure accessor + prefix", ()=> {
-    this.createWrapper(true, ["title"], ["prefix"])
+    this.createWrapper(true, ["title"], ["prefix"], {queryBuilder:QueryString})
 
     expect(this.accessor.key).toBe("q")
     let options = this.accessor.options
     expect(options).toEqual({
-      "queryFields": ["title"],
+      queryFields: ["title"],
       prefixQueryFields:["prefix"],
-      "queryOptions": {}
+      queryOptions: {minimum_should_match:"60%"},
+      prefixQueryOptions: {minimum_should_match:"70%"},
+      queryBuilder:QueryString,
+      onQueryStateChange:jasmine.any(Function)
     })
 
   })
@@ -134,11 +151,11 @@ describe("Searchbox tests", () => {
         .hasClass("is-focused")
     ).toBe(false)
     expect(this.wrapper.node.state)
-      .toEqual({ focused:false })
+      .toEqual({ focused:false, input: undefined })
     this.wrapper.find(".sk-search-box__text")
       .simulate("focus")
     expect(this.wrapper.node.state)
-      .toEqual({ focused:true })
+      .toEqual({ focused:true, input: undefined })
     this.wrapper.update()
     expect(
       this.wrapper.find(".sk-search-box")
@@ -147,5 +164,53 @@ describe("Searchbox tests", () => {
 
   })
 
+
+  describe("url change + blurAction", ()=> {
+
+    it("blurAction:restore", ()=> {
+      this.createWrapper(false, ["title"], ["prefix"], {
+        blurAction:"restore"
+      })
+      this.typeSearch("la")
+      expect(this.wrapper.node.getValue() ).toEqual("la")
+      this.accessor.fromQueryObject({
+        q:"foo"
+      })
+      expect(this.wrapper.node.getValue() ).toEqual("foo")
+
+      this.typeSearch("bar")
+      expect(this.wrapper.node.getValue()).toEqual("bar")
+      this.wrapper.find(".sk-search-box__text")
+        .simulate("blur")
+
+      // should be restored to previous value
+      expect(this.wrapper.node.getValue()).toEqual("foo")
+      expect(this.searchkit.performSearch).not.toHaveBeenCalled()
+
+    })
+
+    it("blurAction:search", ()=> {
+      this.createWrapper(false, ["title"], ["prefix"], {
+        blurAction:"search"
+      })
+      this.typeSearch("la")
+      expect(this.wrapper.node.getValue() ).toEqual("la")
+      this.accessor.fromQueryObject({
+        q:"foo"
+      })
+      expect(this.wrapper.node.getValue() ).toEqual("foo")
+
+      this.typeSearch("bar")
+      expect(this.wrapper.node.getValue()).toEqual("bar")
+      this.wrapper.find(".sk-search-box__text")
+        .simulate("blur")
+
+      // should flush value + search
+      expect(this.wrapper.node.getValue()).toEqual("bar")
+      expect(this.searchkit.performSearch).toHaveBeenCalled()
+
+    })
+
+  })
 
 });