+++ src/components/ui/list/ItemListComponents.tsx
@@ -2,6 +2,7 @@ import * as React from "react";
 
 import { ItemComponent, CheckboxItemComponent } from "./ItemComponents"
 import { ListProps } from "./ListProps"
+import { PureRender } from "../../../core/react/pure-render"
 
 const block = require('bem-cn')
 
@@ -23,6 +24,7 @@ const decodeHtmlEntity = function(str) {
   }).replace(/&amp;/g, '&').replace(/&quot;/g, '"');
 };
 
+@PureRender
 export class AbstractItemList extends React.Component<ItemListProps, {}> {
   static defaultProps: any = {
     mod: "sk-item-list",
@@ -31,6 +33,7 @@ export class AbstractItemList extends React.Component<ItemListProps, {}> {
     translate:identity,
     multiselect: true,
     selectItems: [],
+    countFormatter:identity
   }
 
   isActive(option){
@@ -46,7 +49,7 @@ export class AbstractItemList extends React.Component<ItemListProps, {}> {
   render() {
     const {
       mod, itemComponent, items, selectedItems = [], translate,
-      toggleItem, setItems, multiselect,
+      toggleItem, setItems, multiselect, countFormatter,
       disabled, showCount, className, docCount
     } = this.props
 
@@ -65,7 +68,8 @@ export class AbstractItemList extends React.Component<ItemListProps, {}> {
         bemBlocks: bemBlocks,
         key: option.key,
         itemKey:option.key,
-        count: option.doc_count,
+        count: countFormatter(option.doc_count),
+        rawCount:option.doc_count,
         listDocCount: docCount,
         disabled:option.disabled,
         showCount,