+++ src/core/accessors/Accessor.ts
@@ -2,6 +2,7 @@ import {ImmutableQuery} from "../query/ImmutableQuery";
 import {SearchkitManager} from "../SearchkitManager";
 import {Utils} from "../support"
 const get = require("lodash/get")
+const compact = require("lodash/compact")
 
 export class Accessor {
   searchkit:SearchkitManager
@@ -53,10 +54,13 @@ export class Accessor {
 
   getAggregations(path, defaultValue){
     const results = this.getResults()
-    const getPath = ['aggregations',...path]
+    const getPath = compact(['aggregations',...path])
     return get(results, getPath, defaultValue)
   }
 
+  beforeBuildQuery(){
+
+  }
   buildSharedQuery(query:ImmutableQuery){
     return query
   }