+++ src/core/accessors/QueryAccessor.ts
@@ -12,11 +12,13 @@ const forEach = require("lodash/forEach")
 
 export interface SearchOptions {
   queryFields?:Array<string>
-  prefixQueryFields?:Array<string>
   queryOptions?:any
-  prefixQueryOptions?:any
+  prefixQueryFields?:Array<string>
+  prefixQueryOptions?:Object
   title?: string
   addToFilters?:boolean
+  queryBuilder?:Function
+  onQueryStateChange?:Function
 }
 export class QueryAccessor extends BaseQueryAccessor {
   options:SearchOptions
@@ -26,12 +28,19 @@ export class QueryAccessor extends BaseQueryAccessor {
     this.options = options
     this.options.queryFields = this.options.queryFields || ["_all"]
   }
+  
+  fromQueryObject(ob){
+    super.fromQueryObject(ob)
+    if (this.options.onQueryStateChange){
+      this.options.onQueryStateChange()
+    }
+  }
 
   buildSharedQuery(query){
     let queryStr = this.state.getValue()
     if(queryStr){
-
-      let simpleQuery = SimpleQueryString(
+      let queryBuilder  = this.options.queryBuilder || SimpleQueryString
+      let simpleQuery = queryBuilder(
         queryStr, assign(
           {fields:this.options.queryFields},
           this.options.queryOptions